Refactor previous response serializer service uses with the current format
