+++ src/module/iam/authentication/__test__/authentication.e2e.spec.ts
@@ -7,11 +7,12 @@ import { HttpMethod } from '@common/base/application/enum/http-method.enum';
 
 import { setupApp } from '@config/app.config';
 
-import { SignUpDto } from '@module/iam/authentication/application/service/dto/sign-up.dto';
+import { SignUpDto } from '@module/iam/authentication/application/dto/sign-up.dto';
 import { PASSWORD_VALIDATION_ERROR } from '@module/iam/authentication/infrastructure/cognito/exception/cognito-exception-messages';
 import { CouldNotSignUpException } from '@module/iam/authentication/infrastructure/cognito/exception/could-not-sign-up.exception';
 import { PasswordValidationException } from '@module/iam/authentication/infrastructure/cognito/exception/password-validation.exception';
 import { AppRole } from '@module/iam/authorization/domain/app-role.enum';
+import { User } from '@module/iam/user/domain/user.entity';
 
 import {
   identityProviderServiceMock,
@@ -56,24 +57,23 @@ describe('Authentication Module', () => {
           expect(body).toEqual(
             expect.objectContaining({
               data: expect.objectContaining({
-                email: signUpDto.email,
-                externalId,
-                firstName: signUpDto.firstName,
-                lastName: signUpDto.lastName,
-                avatarUrl: signUpDto.avatarUrl,
-                role: AppRole.Regular,
-                isVerified: false,
+                type: User.getEntityName(),
+                id: expect.any(String),
+                attributes: expect.objectContaining({
+                  email: signUpDto.email,
+                  externalId,
+                  firstName: signUpDto.firstName,
+                  lastName: signUpDto.lastName,
+                  avatarUrl: signUpDto.avatarUrl,
+                  role: AppRole.Regular,
+                  isVerified: false,
+                }),
               }),
               links: expect.objectContaining({
                 self: expect.objectContaining({
-                  href: expect.stringContaining('/user/me'),
+                  href: expect.stringContaining('/auth/sign-up'),
                   rel: 'self',
                   method: HttpMethod.GET,
-                }),
-                update: expect.objectContaining({
-                  href: expect.stringContaining('/user/me'),
-                  rel: 'update',
-                  method: HttpMethod.PUT,
                 }),
               }),
             }),
@@ -112,24 +112,23 @@ describe('Authentication Module', () => {
           expect(body).toEqual(
             expect.objectContaining({
               data: expect.objectContaining({
-                email: signUpDto.email,
-                externalId,
-                firstName: signUpDto.firstName,
-                lastName: signUpDto.lastName,
-                avatarUrl: signUpDto.avatarUrl,
-                role: AppRole.Regular,
-                isVerified: false,
+                type: User.getEntityName(),
+                id: expect.any(String),
+                attributes: expect.objectContaining({
+                  email: signUpDto.email,
+                  externalId,
+                  firstName: signUpDto.firstName,
+                  lastName: signUpDto.lastName,
+                  avatarUrl: signUpDto.avatarUrl,
+                  role: AppRole.Regular,
+                  isVerified: false,
+                }),
               }),
               links: expect.objectContaining({
                 self: expect.objectContaining({
-                  href: expect.stringContaining('/user/me'),
+                  href: expect.stringContaining('/auth/sign-up'),
                   rel: 'self',
                   method: HttpMethod.GET,
-                }),
-                update: expect.objectContaining({
-                  href: expect.stringContaining('/user/me'),
-                  rel: 'update',
-                  method: HttpMethod.PUT,
                 }),
               }),
             }),
@@ -158,24 +157,22 @@ describe('Authentication Module', () => {
           expect(body).toEqual(
             expect.objectContaining({
               data: expect.objectContaining({
-                email: signUpDto.email,
-                externalId,
-                firstName: signUpDto.firstName,
-                lastName: signUpDto.lastName,
-                avatarUrl: null,
-                role: AppRole.Regular,
-                isVerified: false,
+                type: User.getEntityName(),
+                id: expect.any(String),
+                attributes: expect.objectContaining({
+                  email: signUpDto.email,
+                  externalId,
+                  firstName: signUpDto.firstName,
+                  lastName: signUpDto.lastName,
+                  role: AppRole.Regular,
+                  isVerified: false,
+                }),
               }),
               links: expect.objectContaining({
                 self: expect.objectContaining({
-                  href: expect.stringContaining('/user/me'),
+                  href: expect.stringContaining('/auth/sign-up'),
                   rel: 'self',
                   method: HttpMethod.GET,
-                }),
-                update: expect.objectContaining({
-                  href: expect.stringContaining('/user/me'),
-                  rel: 'update',
-                  method: HttpMethod.PUT,
                 }),
               }),
             }),
