+++ src/module/iam/user/application/dto/user-response.dto.ts
@@ -3,7 +3,6 @@ import { IDto } from '@common/base/application/dto/dto.interface';
 import { AppRole } from '@module/iam/authorization/domain/app-role.enum';
 
 export class UserResponseDto implements IDto {
-  id: string;
   email: string;
   firstName: string;
   lastName: string;
@@ -13,7 +12,6 @@ export class UserResponseDto implements IDto {
   avatarUrl?: string;
 
   constructor(
-    id: string,
     email: string,
     firstName: string,
     lastName: string,
@@ -22,7 +20,6 @@ export class UserResponseDto implements IDto {
     avatarUrl?: string,
     externalId?: string,
   ) {
-    this.id = id;
     this.externalId = externalId;
     this.email = email;
     this.firstName = firstName;
