+++ src/module/iam/user/user.module.ts
@@ -3,7 +3,6 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { UserMapper } from '@module/iam/user/application/mapper/user.mapper';
 import { USER_REPOSITORY_KEY } from '@module/iam/user/application/repository/user.repository.interface';
-import { UserLinkBuilderService } from '@module/iam/user/application/service/link-builder/user-link-builder.service';
 import { UserPostgresRepository } from '@module/iam/user/infrastructure/database/user.postgres.repository';
 import { UserSchema } from '@module/iam/user/infrastructure/database/user.schema';
 
@@ -14,7 +13,7 @@ const userRepositoryProvider: Provider = {
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserSchema])],
-  providers: [userRepositoryProvider, UserMapper, UserLinkBuilderService],
-  exports: [userRepositoryProvider, UserMapper, UserLinkBuilderService],
+  providers: [userRepositoryProvider, UserMapper],
+  exports: [userRepositoryProvider, UserMapper],
 })
 export class UserModule {}
